Separate out round and category fulfillment logic.
