+++ functions/src/fulfillments/__tests__/categoryFulfillment.test.ts
@@ -1,7 +1,7 @@
 import { ConversationData, ResponseType } from '../../models/models';
-import { selectRound, startCategory } from '../categoryFulfillment';
 
 import { Topic } from '../../models/rounds';
+import { startCategory } from '../categoryFulfillment';
 
 describe('Start Category', () => {
   test('Topic and question number are set correctly if not on ConversationData', () => {
@@ -48,32 +48,5 @@ describe('Start Category', () => {
     };
     const response = startCategory(topicChoice, data);
     expect(response.responseType).toEqual(ResponseType.ASK);
-  });
-});
-
-describe('Select Round', () => {
-  test('If no round set round to 1', () => {
-    const data: ConversationData = {
-      startRepromptIssued: false,
-    };
-    const expectedData: ConversationData = {
-      startRepromptIssued: false,
-      currentRound: 1,
-    };
-    selectRound(data);
-    expect(data).toEqual(expectedData);
-  });
-
-  test('Update round correctly if round exists on ConversationData', () => {
-    const data: ConversationData = {
-      startRepromptIssued: false,
-      currentRound: 3,
-    };
-    const expectedData: ConversationData = {
-      startRepromptIssued: false,
-      currentRound: 4,
-    };
-    selectRound(data);
-    expect(data).toEqual(expectedData);
   });
 });
